server intialized
